requirements.txt
PyYAML >= 5.4
plotly >= 5.0
pandas >= 1.0
textblob >= 0.15.3
scikit-learn >= 0.24
